requirements.txt
appnope==0.1.0
certifi==2017.4.17
decorator==4.0.11
Django==1.11
djangorestframework==3.6.2
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
